store parameter style in prepared statement
