+++ statement/statement.go
@@ -3,35 +3,44 @@ package statement
 import (
 	"context"
 	"database/sql/driver"
+
+	contextUtils "github.com/firebolt-db/firebolt-go-sdk/context"
 )
 
-type fireboltStmt struct {
-	execer  driver.ExecerContext
-	queryer driver.QueryerContext
-
-	query string
+type Executor interface {
+	ExecutePreparedQueries(ctx context.Context, queries []PreparedQuery, args []driver.NamedValue, isQuery bool) (driver.Rows, error)
 }
 
-func MakeStmt(execer driver.ExecerContext, queryer driver.QueryerContext, query string) (*fireboltStmt, error) {
+type fireboltStmt struct {
+	Queries  []PreparedQuery
+	executor Executor
+}
+
+func MakeStmt(executor Executor, query string, style contextUtils.PreparedStatementsStyle) (*fireboltStmt, error) {
+	preparedQueries, err := prepareQuery(query, style)
+	if err != nil {
+		return nil, err
+	}
+
 	return &fireboltStmt{
-		execer:  execer,
-		queryer: queryer,
-		query:   query,
+		Queries:  preparedQueries,
+		executor: executor,
 	}, nil
 }
 
 // Close the statement makes it unusable anymore
 func (stmt *fireboltStmt) Close() error {
-	stmt.execer = nil
-	stmt.queryer = nil
-
-	stmt.query = ""
+	stmt.executor = nil
+	stmt.Queries = nil
 	return nil
 }
 
-// NumInput returns -1, parametrized queries are not implemented
+// NumInput returns the number of input parameters
 func (stmt *fireboltStmt) NumInput() int {
-	return -1
+	if len(stmt.Queries) == 0 {
+		return -1
+	}
+	return stmt.Queries[0].GetNumParams()
 }
 
 // Exec calls ExecContext with dummy context
@@ -46,10 +55,14 @@ func (stmt *fireboltStmt) Query(args []driver.Value) (driver.Rows, error) {
 
 // QueryContext sends the query to the engine and returns fireboltRows
 func (stmt *fireboltStmt) QueryContext(ctx context.Context, args []driver.NamedValue) (driver.Rows, error) {
-	return stmt.queryer.QueryContext(ctx, stmt.query, args)
+	return stmt.executor.ExecutePreparedQueries(ctx, stmt.Queries, args, true)
 }
 
 // ExecContext sends the query to the engine and returns empty fireboltResult
 func (stmt *fireboltStmt) ExecContext(ctx context.Context, args []driver.NamedValue) (driver.Result, error) {
-	return stmt.execer.ExecContext(ctx, stmt.query, args)
+	_, err := stmt.executor.ExecutePreparedQueries(ctx, stmt.Queries, args, true)
+	if err != nil {
+		return nil, err
+	}
+	return FireboltResult{}, nil
 }
